feat: expand city list and tweak sun markers
